changed CreateAppointmentViewModel AppointmentLength to double

+++ ViewModels/CreateAppointmentViewModel.cs
@@ -1,4 +1,3 @@
-
 using System;
 
 namespace tutorfy_backend.ViewModels
@@ -6,7 +5,7 @@ namespace tutorfy_backend.ViewModels
     public class CreateAppointmentViewModel
     {
         public DateTime StartTime { get; set; }
-        public int AppointmentLength { get; set; }
+        public double AppointmentLength { get; set; }
         public string Location { get; set; }
     }
 }
